Rid of the intermediate internal package as 7.0.x branch does not require external tools to generate package configurations
